config/requirements.txt


django 
djangorestframework
django-cors-headers 
Pillow
djangorestframework-simplejwt